Working through `size` (2).

+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/datetime/DateStorage.java
@@ -64,7 +64,7 @@ public final class DateStorage extends SpecializedStorage<LocalDate> {
     t.add(
         new TimeLikeCoalescingOperation<>(Maps.MIN, LocalDate.class) {
           @Override
-          protected Builder createOutputBuilder(int size) {
+          protected Builder createOutputBuilder(long size) {
             return Builder.getForType(DateType.INSTANCE, size, null);
           }
 
@@ -76,7 +76,7 @@ public final class DateStorage extends SpecializedStorage<LocalDate> {
     t.add(
         new TimeLikeCoalescingOperation<>(Maps.MAX, LocalDate.class) {
           @Override
-          protected Builder createOutputBuilder(int size) {
+          protected Builder createOutputBuilder(long size) {
             return Builder.getForType(DateType.INSTANCE, size, null);
           }
